requirements.txt
python-telegram-bot==22.7
boto3==1.34.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
cryptography==41.0.7
apscheduler==3.10.4
